[FMD-103] Override work_mem for known slow queries
Mostly these are queries which export a lot of data. Note that though
this setting can be set on the whole cluster, there is plenty of advice
not to set it at that scope and instead to use either transaction or
database level. In this case we set it on the txn level.

See the section titled _Considerations_ in
https://aws.amazon.com/blogs/database/tune-sorting-operations-in-postgresql-with-work_mem/