build assets in python script
Moves the downnloading of the govuk-frontend assets and building bundled assets into a standalone script that can be run at buildtime.

We still support dynamically updating the static assets in development environment, but need to disable this in deployed environment so we are not writing to the deployed instances filesystem.