fix: ensure end dates cannot be earlier than start dates
We have a number of models that have start/end date pairs, but no
constraint that ensures end dates cannot be earlier than start dates.

This patch adds those constraints via SQL CHECK CONSTRAINTs, and fixes
some test data that broke assurances.

We will release the DB change in two migrations:
1) add the constraints as NOT VALID
2) validate the constraints

After migration 1, the constraints will be enforced for all new
transactions but existing data has not been checked. This avoids doing a
full table scan which would take an `ACCESS EXCLUSIVE` lock.

To validate the constraint in migration 2, we then only need to take a
SHARE UPDATE EXCLUSIVE lock on the un-validated rows. This allows
reading, inserting new rows, and updating validated rows at the same
time. We are *NOT* including that constraint validation step in this PR,
as we know that it will fail because some current data does not respect
the constraint.

We will fix the data and merge the VALIDATE CONSTRAINT step separately,
once we know it will work.