fix: ignore na elements when running manual checks
Element-wise pandera checks receive data in whatever type pandas has
inferred is most appropriate. If just numeric data, an int or float, if
a datetime string, then a datetime python object, etc. This is not the
type that we coerce the column to eventually via pandera.

If the cell is empty, pandas/pandera passes us a `float('nan')` object.
At the moment most of our checks will reject this explicitly as
inappropriate by doing an `isinstance` check and raising a TypeError.

Our validation logic discards failed checks due to TypeErrors, as these
are expected to be issues around dtype conversion, and we report on
those errors separately. However, raising a TypeError from a pandera
check also eagerly terminates validation on the rest of the rows of the
column, so we potentially skip other error information. Let's check for
missing data and explicitly fail the check. Panderas has some specific
logic for this case, as it disattributes the error related to the check
if the cell is empty. So if we are validating postcodes, an empty cell
can fail the 'is the format of the postcode valid' check without
flagging an error saying 'you entered an invalid postcode'. They will
only get the separate error for 'blank data in cell'.