fix: fail rather than error on wrong postcode dtype
When running the postcode_list check, the element passed in via
pandas/pandera can be of any type: str, int, float, datetime, etc.

If it's not a string, we currently raise a TypeError. But our table
validation discards type errors, relying on data type conversion later
in the pipeline logging the appropriate error. But that only happens if
it's not possible to convert this data to the target datatype.

In the case of postcodes, our target data type is a string. If the
source data type is an int or a datetime, that _can_ be coerced to a
string later and so wouldn't error later in the pipeline.

We need to catch that validation issue up front. So, if the postcode
element under test is anything other than a string, let's fail the check
outright. Not with an exception, but in the standard way, so that
pandera flags the error properly.