Clean and sort Message cell indexes
As it stands, cell indexes for Messages (errors) are concatenated
together in whatever arbitrary order they are created/combined. This can
sometimes lead to a list of cell references like: A1, B2, C3, A2, B1,
C1.

The lack of ordering probably makes this harder for users to parse and
follow, and more likely for them to miss a cell references when
scanning.

Let's sort cell orders deterministically so that they appear in a
reliable and logical order: sorted by column first, then by row number.

Two important things to note:
* Cell indexes can just be column identifiers, with no row number.
* Column identifiers can be multiple letters, and all 1-letter column
  identifiers come before any 2-letter column identifiers, etc.