Implement basic sentry metrics tracking
Add support for tracking custom business metrics for reporting
submission uploads, so that we can easily see some basic stats around
how many submissions are happening, and their results
(success/validation issues/server errors).

Because of the way that the three dimensions we're interested in reach
the API (basically all contained in an excel spreadsheet data blob), we
store the data on `Flask.g`, which is a request-scoped thread-local
store. This means we can safely put data here and access it as,
essentially, global values, but that it won't be visible to other
threads processing simultaneous requests and it will be cleaned up
automatically at the end of each request. This is a bit of a
quick-and-dirty workaround, but restructuring the ingest function to
expose this data more cleanly would require significant additional work
that doesn't seem worth it at this time. We may rip out these metrics
after trialling them, and/or we may not handle submissions via
spreadsheets in the general case for long (at least, here's hoping - see
you in 10 years when I've been proven wrong).