SMD-773 - Short-term fix for unhandled dropdown placeholders in bespoke outputs / outcomes tables
Unhandled dropdown placeholder is e.g., "You have no bespoke outputs to select" for "Bespoke outputs" table. If the user selects this as "Output", then values are expected across the entire row, which is unintuitive for the user. The idea is that this is a short-term fix and can be removed pending a spreadsheet change for Pathfinders R2