SMD-711 - Replacing hard-coded enum checks with dynamic cross-table checks
At the moment there are lists of values that are used in the spreadsheet to populate the dropdowns that the user interacts with, and we effectively replicate those lists on the backend for the purposes of validation. However, this approach means that we have two conflicting sources of truth and can lead to misalignment if not maintained properly.

This change removes the hard-coded checks that are performed at schema validation time, replacing the functionality of each with the addition of corresponding cross-table validation checks. These checks validate data against the values extracted from the hidden worksheet 'Dropdown Values' in the spreadsheet, creating it as a single source of a truth.