fix: close `div` properly for selectItems
Possibly due to inconsistent indenting, there was a missing closing
`div` tag, which could cause HTML rendering issues in certain cases.

In one specific case, on the download page, selectItems was the last
part of an accordion, and it caused the styling for the end of the
accordion to leak beyond the intended range. The accordion inserts a
bottom border line after the each accordion section, and this leak was
causing that border line to appear well beyond where we expected (in
this case, after the download button and 'Access the funding glossary'
link. This only happened after fixing some other HTML issues on the
page.