bau: remove start page
The start page isn't part of the main user journey any more. As soon as
a user logs in or access the root page (`/`), they are redirected to
`/download`.

It looks like a decision was made to skip the start page as being fairly
unnecessary, but the code/view wasn't ever removed. So let's tidy it up
now.

Just in case any users have it bookmarked, we add a redirect so that
things don't break for them.