SMD-819 - Using deepcopy to avoid mutating fixture mock_df_dict
This is a temporary solution to keep the scope of the PR limited to the ticket. The plan is to make the fixture function-scoped and to adjust test methods containing mutation accordingly. Function-scoping the fixture will only work if we define the mock data fresh every time, and do not simply recycle a reference to pre-defined data.