Update alembic migration file format
We want our migration files to follow a consistent format - the 3-digit
revision ID followed by a short description.

Postgres identifiers are limited to 32 characters, so let's truncate the
message part at 28 characters. With the 3 characters of the slug and the
1 character of the underscore this lets us max out the information in
the version id.

Alembic enforces this restriction by setting the `version_num` column as
a `varchar(32)`. It's not configurable.