SMD-842 - Moving 'tables' under 'core' directory and moving 'tables/tests' under 'tests' directory
Code under 'tables' can definitely be considered part of the core application at this point. I believe Cyrus originally created a separate directory outside of 'core' to house this stuff just to keep it isolated during development, but it's time to move it where it belongs, to improve the logical layout of the repo as a whole.