SMD-503 - Tie Pandera check functions to error messages
Currently we register custom Pandera check functions (used in schema validation for Pathfinders) through Pandera's register_check_method decorator. The problem is that we cannot specify an error message that should be raised if the check fails, and that means that every time we instantiate a check in our code, we have to remember to pass the error message that corresponds to the check as a separate argument, which is error-prone and verbose.

As check functions are very much tied to error messages in the real world, for the sake of a consistent user experience, it makes sense to tie them together formally in the backend, instead of simply by convention. Thus, this change is to remove the practice of registering check functions via a decorator and handing over the process of actually instantiating these checks to Pandera, and instead instantiate checks in our own application code, facilitating the passing of error messages as arguments. As we also need to support arguments, we cannot instantiate them directly in the 'checks.py' module and instead need to offer wrapper functions or closures.

To facilitate this tying together of check functions and error messages for all checks used during Pandera schema validation for Pathfinders, checks that do not use custom logic have also been assigned closures in the 'checks.py' module (e.g., the email regex check, which uses Pandera's built-in str_matches check function).

As checks are now tied to error messages, tests had to be updated to reflect this and ensure that the real error messages (from PFErrors) were being verified against.

Finally, the check function exactly_five_rows was replaced by the more generic exactly_x_rows, slightly out of scope but definitely the right thing to do.