SMD-503 - Moving const classes from config module to new consts module under core/validation
The const / enum classes PFErrors, PFRegex and PFEnums belong in a separate module to the round-specific config module they are currently found in. All are specifically tied to the schema validation process, for which we already have modules including checks.py and exceptions.py under core/validation. A new consts.py should be created in the same location as those other modules to house these const / enum classes.