SMD-503 - PF validation - tying together check functions and error messages for existing custom checks
Currently we register custom Pandera check functions (used in schema validation for Pathfinders) through Pandera's register_check_method decorator. The problem is that we cannot specify an error message that should be raised if the check fails, and that means that every time we instantiate a check in our code, we have to remember to pass the error message that corresponds to the check as a separate argument, which is error-prone and verbose.

As check functions are very much tied to error messages in the real world, for the sake of a consistent user experience, it makes sense to tie them together formally in the backend, instead of simply by convention. Thus, this change is to remove the practice of registering check functions via a decorator and handing over the process of actually instantiating these checks to Pandera, and instead instantiate checks in our own application code, facilitating the passing of error messages as arguments. As we also need to support arguments, we cannot instantiate them directly in the 'checks.py' module and instead need to offer wrapper functions or closures.

This commit specifically targets only checks that have existing custom check functions.