SMD-653 - Dependency injection for Pathfinders
Currently we have an IngestDependencies class that is used by Pathfinders but only for initial validation schema and table to load function mapping, as the other dependencies don't map directly between Towns Fund and Pathfinders. For the dependencies that Towns Fund has but not Pathfinders, we specify None in the constructor for the IngestDependencies object when doing a Pathfinders run, while for the dependencies that Pathfinders has but not Towns Fund, we hard-code them in the main 'ingest' pipeline flow.

This change is to break apart IngestDependencies into three classes - a base IngestDependencies, containing shared dependencies; and then specific child classes for Towns Fund and Pathfinders, containing fund-specific dependencies. This allows us to share the same dependency factory method but to return IngestDependency objects specific to funds. We can then leverage this change to replace the hard-coding of Pathfinders-specific dependencies with dependency injection, similar to what Towns Fund enjoys now, which will help when we move to Pathfinders round 2.