Stop reading `reporting_round` from submissions
Now that we have the `reporting_round` tracked on the
`programme_junction` table, lets switch all reads from
`submission.reporting_round` over to
`programme_junction.reporting_round`.

As well as changing DB reads, we also stop reading the `Submission_Ref`
dataframe that eventually populates the `submission` table. Instead, we
rely on the `reporting_round` variable passed into `ingest`, and we
cascade this through to `populate_db` so that it can avoid checking the
`Submission_Ref` data frame. This is because we'll shortly stop writing
`submission.reporting_round`, and remove it from the data frame, so
we're just preparing for that.