Integrate vite to bundle gov.uk frontend css+js
Vite appears to be the current bundler of the moment. Using this will
let us extend and modify GOV.UK Frontend styles so that we can easily
build on the work done with any future components we need. It also lets
us configure some option, ie to disable use of GDS Transport.

We need to add routes to flask to serve static assets from all
subdomains.

Removes flask-assets, cssmin, and jsmin, which are now handled by vite
via npm