Switch to serving vite assets statically
When using the 'development' mode, I see the styles 'pop' in - the HTML
is rendered unstyled and then the styles are applied a bit later. It
looks gross. Pulling the statics in directly looks better.

The downside here is that the 'development' supposedly gives
hot-reloading of css/js changes. If we need that, maybe we can switch
back. For now I don't care.