Improve TF validation
Return user friendly error for cells that are expected to be str but
datetime is provided.

Add generic message for type mismatch, however to not break existing
currency validation, we ignore numbers.

Especially for Funding Question section. Where the `Response` column
must be mapped to a particular `Question` sheet column.