Fix migration
Pathfinders funding sometimes has no funding_source_name in data_blob.
We set it to `Pathfinders` when that's the case, additionally to ensure
we don't touch other fundings, we also check for `project_id` to be
null.

Update the removal of jsonb properties, as the syntax was incorrect.

Remove lines that make data_blob nullable at the start of the migration and
not nullable at the end of the migration, as I think they are
unnecessary.