Prevent overlapping form fields
When initialising form submission structures, check that there are no
sections/subsections/pages/form fields that have overlapping keys as a
safety check against overwriting data.

And then fix this for RAG ratings where it was happening.

We could avoid this by storing each page's form data in a separate
dictionary (ie adding the page's path_fragment as another level of
nested keys). I ended up not doing this because it 'feels'/'looks' nicer
to have all of the data for a subsection in one dict. Maybe that's a bad
call though...