Switch from subdomain matching to host matching
`subdomain_matching` and `host_matching` are mutually exclusive options
in Flask. If you set `subdomain_matching`, you must also configure
`SERVER_NAME`, which Flask/Werkzeug then checks for every incoming
request. Any request that doesn't match SERVER_NAME is rejected (404)
outright, and I couldn't find any way to override this for even a single
route. We need to expose at least one route (the healthcheck for ECS) on
a IP:port host, because it comes via the private/internal IP for the ECS
container.

So we need to switch to using `host_matching`, and specifying the host
for all of our routes explicitly. Exploring this option led me to:

https://github.com/pallets/flask/discussions/5413

which discusses some confusion/inability to set a `host` on an entire
blueprint. It only seems possible to do this on an individual route
level by default. In the comments, though, we find a link to:

https://github.com/tachyondecay/tachyondecay.net-mirror/blob/main/lemonade_soapbox/helpers.py#L79-L112

where a workaround/implementation has already been provided to add a
`host` parameter to blueprints so that all routes are registered on that
host. Thank you open source MIT licenced repos.

So this patch switches us over to host_matching.

An extra 'gotcha' with host_matching is that, if you're serving on
something other than 80 or 443, the port needs to be included in the
host. So for local development (and for the ALB healthcheck), our 'host'
will be something like '127.0.0.1:4001', or '10.0.0.100:8271' or
whatever. So we need to fiddle around when building the host and make
sure to append the port we're on.