Make healthchecks faster
By default, AWS CoPilot has the following settings:

healthy_threshold: 5
interval: 30s
timeout: 5s
grace_period: 60s

This means that AWS waits for 60 seconds before starting to ping the
service's healthchecks, then it will ping the healthcheck every 30
seconds, and wait up to 5 seconds for a response. After 5 successful
responses, it will consider the service healthy.

This means that a new service will never be considered healthy until at
least 60s + (5 * 30s) have passed, which is 3.5 minutes.

We know our flask app starts fairly quickly, so we don't need such a
long grace period. Our healthcheck is very light-weight, so we should be
fine to ping it fairly often. And then finally I don't see any reason
that our healthcheck would need to pass 5 times to be considered
healthy; we are a low volume service and the healthcheck should be very
reliable, so if the healthcheck passes twice in a row we should be
fairly confident it's fine.

Untouched is `unhealthy_threshold`, which defaults to 2, which will make
AWS consider the instance unhealthy if two healthchecks fail in a row.