Update github deployment pipeline to just deploy data-store
We now have two separate services/workloads to deploy via copilot -
data-store (the API) and data-store-celery (the celery worker).

I've decided to deploy the celery worker first, so that if we merge any
new tasks to the codebase, the worker comes up first and is immediately
available to handle any jobs that get queued by the API.

We need to make sure that any changes we release to celery jobs are
backwards-compatible for at least one release/deployment. This should
generally be our principle anyway, so I think nothing significant
changes here.