Add S3 bucket access to data-store-celery
The celery worker needs to be able to read/write to the S3 bucket that
we store generated find download files in. That bucket is a
workload-level addon for the data-store service, so already exists. We
create an addon file in the data-store-celery workload that just creates
an IAM access policy. By including this policy in the Outputs section,
Copilot automatically assigns it to the ECS task role. We then also
output the bucket name so that Copilot automatically injects that value
under the AWS_S3_BUCKET_FIND_DATA_FILES environment variable.