Prevent downloads from failing on empty strings
When using the swagger UI to hit `trigger_async_download`, if the `send
empty values` checkbox is filled, then the API request sent for the
`organisation` (for example) field will be: `['']`.

This is treated as 'truthy' for the query, so would end up running some
SQL like `WHERE organisation_id = ''::uuid`.

SQL can't convert an empty string to a UUID, and so throws an error.

Let's fix this by filtering out any empty string values first.