Initial playwright setup - test find flow
Adds a browser test using Playwright that walks through the main user
journey for a find user.

This involves:

* selecting some filters
* select a download file format
* clicking download
* clicking the link in the email that is received later
* downloading the final report on the page

We need to do some incredibly funky stuff around authentication to get a
valid user session (in a way that could feasibly work against dev and
test environments, too). We (ab)use authenticator's magic links flow to
create a link as if we're an assessor of a new "post-award-e2e-tests"
fund (which we inject into the local mock data for authenticator in a PR
on that repo). When we use the magic link, it sets a cookie on the
.levellingup.gov.localhost that is valid for Find, so we become
effectively logged in.