Turn warnings into errors so that we always deal with them
At the moment, warnings emitted during the test run are reported at the
end. But we've basically just been ignoring these and not doing anything
about them.

We can instead tell pytest to report them as errors, which will result
in failed test runs. This will force us to deal with warnings emitted by
our (or our dependencies') code. These can be things like
DeprecationWarnings, which can be early warning signs for our
dependencies changing how things are done and that we need to make
changes in our code.

Or they can be telling us that we're using some library incorrectly or
imprecisely, and should probably do something different.

To make sure that we're staying on top of our code, we should have a
rule that all warnings are addressed. This can either be through us
fixing/changing something in our code, or by explicitly ignoring the
warning (to say "we acknowledge this and accept it").