Add exception for missing postcode data
`add_project_geospatial_relationship` is in our `data_store` module,
which should need no concept of HTTP requests. This function, however,
uses flask's `abort` call in situations where an error has happened.
Instead, we should just raise an exception, which will bubble up to the
relevant view and then be handled by flask as a 500. (Or it could be
caught and handled appropriately)