Updating Pathfinders transform pipeline to populate reporting_round table and reporting_round_id columns on submission
The migration handles the population of this table and these columns for data already in the database, but we also need to populate these when a new submission comes in, to ensure that these changes are deployable without the possibility of missing data. A ReportingRound DataFrame is added to the output from Pathfinders transformation, which is then used during database load to either create a new or retrieve an existing reporting round entity from the database. We do this in advance of data mapping, so that the ID is available to the submission and programme_junction tables. The alternative - leveraging the existing foreign key mapping process - would have been very messy, as we'd have needed to store fund code and round number on the transformed DataFrame for submission_dim, and then extend map_data_to_models to handle a double look-up, and even then programme_junction would've needed to be handled distinctly via its load function.

As part of this change, we fix Pathfinders-specific tests, but as the main pipeline is shared between Towns Fund and Pathfinders, this causes lots of failing tests elsewhere, which will be fixed in the coming commits.