Updating Towns Fund transform pipelines to populate reporting_round_tables and reporting_round_id columns on submission, and fixing up all tests
In our last commit we fixed up the Pathfinders transform pipeline, this time its the turn of the remaining transform pipelines, which also gives us the change to fix up all remaining failing tests not specific to Pathfinders. The changes here basically involve adding a new ReportingRound DataFrame to the Towns Fund transformed output, as the changes to the shared components necessary were done in the previous commit to ensure Pathfinders-specific tests passed.