Fixing local setup issue
During local setup, seed-ref runs before seed-sample-data. Previously on this branch, we added reporting_round seeding to seed-sample-data to get tests passing, but because reporting_round is also seeded during seed-ref, this ends up causing problems when you try to run data-store locally with Docker-Compose, as it is trying to insert a row that already exists. The solution is to remove reporting_round seeding from seed-sample-data and instead seed it in tests in the same way that other reference data is seeded, getting tests to pass and also allowing data-store to boot up successfully locally.