FPASF-436 - Create reporting round table and populate for historic and new submissions (#668)
* ReportingRound SQLAlchemy entity and associated changes

Associated changes include new columns reporting_round_id on the Submission and ProgrammeJunction entities, and new relationships defined between ReportingRound and Fund, Submission and ProgrammeJunction.

* Migration to create reporting_round table and new columns and foreign key constraints on submission_dim and programme_junction tables

* Updating Pathfinders transform pipeline to populate reporting_round table and reporting_round_id columns on submission

The migration handles the population of this table and these columns for data already in the database, but we also need to populate these when a new submission comes in, to ensure that these changes are deployable without the possibility of missing data. A ReportingRound DataFrame is added to the output from Pathfinders transformation, which is then used during database load to either create a new or retrieve an existing reporting round entity from the database. We do this in advance of data mapping, so that the ID is available to the submission and programme_junction tables. The alternative - leveraging the existing foreign key mapping process - would have been very messy, as we'd have needed to store fund code and round number on the transformed DataFrame for submission_dim, and then extend map_data_to_models to handle a double look-up, and even then programme_junction would've needed to be handled distinctly via its load function.

As part of this change, we fix Pathfinders-specific tests, but as the main pipeline is shared between Towns Fund and Pathfinders, this causes lots of failing tests elsewhere, which will be fixed in the coming commits.

* Updating Towns Fund transform pipelines to populate reporting_round_tables and reporting_round_id columns on submission, and fixing up all tests

In our last commit we fixed up the Pathfinders transform pipeline, this time its the turn of the remaining transform pipelines, which also gives us the change to fix up all remaining failing tests not specific to Pathfinders. The changes here basically involve adding a new ReportingRound DataFrame to the Towns Fund transformed output, as the changes to the shared components necessary were done in the previous commit to ensure Pathfinders-specific tests passed.

* Renaming reporting_round and fund_id variables to round_number and fund_code for clarity

* Refactoring Towns Fund transform pipelines to use fund_code calculated in parent function

* Seeding reporting round data

* Updating .current-alembic-head file to point to latest migration

* Fixing local setup issue

During local setup, seed-ref runs before seed-sample-data. Previously on this branch, we added reporting_round seeding to seed-sample-data to get tests passing, but because reporting_round is also seeded during seed-ref, this ends up causing problems when you try to run data-store locally with Docker-Compose, as it is trying to insert a row that already exists. The solution is to remove reporting_round seeding from seed-sample-data and instead seed it in tests in the same way that other reference data is seeded, getting tests to pass and also allowing data-store to boot up successfully locally.

* Ensuring that Find continues to work in light of change to reporting period end datetime

---------

Co-authored-by: Gideon Goldberg <1764158+gidsg@users.noreply.github.com>