Migrate DB before code deployments
Our current deployment pipeline runs three jobs in parallel:

1. Migrate DB and deploy post-award service
2. Deploy post-award-celery service
3. Deploy download-report job

If the DB migration fails, we still deploy post-award-celery and the
download-report job. These might be relying on model changes from a DB
migration, so it is logically incorrect to be deploying these if the
migration fails. It also leaves us in an inconsistent place - post-award
service is on the old code, post-award-celery/download-report are on the
new code.

We should run the DB migrations before deploying any of the code
changes. If these succeed, we can run all three deployments. If they
fail, we shouldn't run any.