Allow e2e tests to target dev+env environments
This adds the '--e2e-env' option to our e2e tests so that they can
target the deployed dev and test environments.

By default these will still just run against the local domains, but this
is a step towards us being able to run these tests in CI.

For now, the find test remains broken in dev+test envs because we are
hardcoding our filter choices, and these are not reliably present in
each environment. This will be fixed later, but getting this in now
should unblock another developer's work on playwright tests.