Make end-to-end tests work against dev/test envs
Update find test to work on dev and test envs

We don't have reliable values for each of these filter values, so let's
read the available values and pick some random ones.

---

Centralise access to e2e test config

We are currently pulling a few config variables from the `UnitTest`
config object - but this doesn't work well for a few reasons:

- It pulls from environment variables, so switching between
  local/dev/test environments means toggling a bunch of variables
  manually.
- Injecting those variables in CICD means duplicating the secrets in
  GitHub, rather than pulling direct from the source (ie AWS).

If we have a separate store for the end-to-end test config, we can add
extra logic to simplify retrieving the correct values for the correct
local environment (developer laptop or github cicd), and target
environment (developer laptop, or deployed AWS environment).

This is a stub that will be added to in the following commits.

---

Change basic auth flow

Instead of baking the basic auth credentials into the domain, we inject
them into the playwright browser context. This is so that the browser
caches them, which will let them work when we try to hit a dev/test URL
directly that doesn't have the credentials injected (eg the link in
emails that takes a user to their find report for downloading).

---

Change Notify API key flow for e2e tests

Rather than using environment variables to set the Notify API keys in
all cases, only use this when running against a local developer
environment.

When targetting AWS (dev or test), we will pull these credentials from
AWS (because they need to be scoped the same as the deployed
environment, and because we want to be able to reliably access them in
eg github actions without having to duplicate the secret).

---

Build JWT directly rather than (ab)use auth magic links

We are switching from, really, one hack to another here, in order to be
able to run these tests against and in the dev/test environments.

Currently/previously, we used the magic links flow on authenticator to
get a logged-in session, but this only worked easily for accounts that
don't need any roles. To give them roles, we need to hit the account
store API directly to inject them. The account store API is internal
(not accessible over the public Internet) in dev/test environments, so
we would need to do some janky stuff to get access to it from eg local
machines and/or github actions.

Our current auth flow just ends up setting a cookie in the browser that
works across all FS domains and doesn't ever "call home" to verify the
details. So if we can just build a matching cookie and inject it into
the browser, those details will be accepted outright. We need to sign it
with a private key that the apps will trust, so we now need to read that
key appropriately for the dev/test environments. Then just build the
JSON blob, sign it, and stick it in the browser.

---

- Remove unused authenticator playwright pages

- Add github action for e2e tests

- Update README instructions for e2e tests