Move from script hashes to CSP nonces
Previously, we allowed custom JS to execute in the browser by
whitelisting the hashes of the script contents. This had a few issues

* We reference some JS by domain names, so needed to have the hash for
 each of our environments whitelisted.
* We didn't comment these well, so it's unclear which hashes are
  whitelisting which scripts.

We now stop using the hash whitelist and move over to CSP nonces. These
are random values that we associate with each custom inline script to
tell the browser they're trusted and can be executed.

The CSP nonce present on each script tag must match the CSP nonce passed
back in the HTTP response header, otherwise the browser will refuse to
run the script.