Report browser console errors
Playwright lets you read messages sent to the JavaScript console, as
well as the level they're written at (eg info, warning, error).

Let's capture errors and raise those in pytest as warnings, which will
get caught and report as test failures.

This should help us catch anything broken client-side, for example if a
script fails to load because of a CSP violation.