Use ReportingRound as reference data
Stop using the hard-coded dictionaries that contain reporting round
dates. Instead, we can now just reference the reporting round records
in the DB directly.

This change means that ingesting a spreadsheet will no longer *create* a
new ReportingRound entry - this is now a thing that needs to be done
manually by a developer as part of setting up a new reporting round for
a fund. It's very low maintenance, however, using the flask-admin
interface.